requirements.txt
cffi               
pycparser
pycurl
click
enum-compat
eventlet
Flask==1.0.4
Flask-SocketIO
gevent
gevent-websocket
greenlet
gunicorn
Jinja2
MarkupSafe
numpy==1.16.0
olefile
opencv-python==3.4.11.43
Pillow==4.2.1
python-engineio
python-socketio
six
glob3
face_recognition
keras==2.0.8
TensorFlow==1.13.2

